requirements.txt
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
Brlapi==0.6.4
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
checkbox-support==0.22
chrome-gnome-shell==0.0.0
click==6.7
command-not-found==0.3
constantly==15.1.0
cryptography==2.2.2
cycler==0.10.0
defer==1.0.6
Django==2.0.4
dlib==19.10.0
docopt==0.6.2
face-recognition==1.2.2
face-recognition-models==0.3.0
feedparser==5.1.3
Flask==1.0.2
future==0.16.0
gdbn==0.1
gnumpy==0.2
guacamole==0.9.2
html5lib==0.9999999
httplib2==0.9.1
hyperlink==18.0.0
idna==2.6
ImageHash==4.0
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
joblib==0.11
Keras==2.0.8
kiwisolver==1.0.1
language-selector==0.1
Lasagne==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.2.2
nolearn==0.5b1
nose==1.3.7
numpy==1.14.2
oauthlib==1.0.3
onboard==1.2.0
opencv-python==3.4.0.12
padme==1.1.1
pandas==0.17.1
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
protobuf==3.4.0
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.2.1
pycparser==2.18
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyOpenSSL==17.5.0
pyparsing==2.0.3
pyTelegramBotAPI==3.5.1
python-apt==1.1.0b1+ubuntu0.16.4.1
python-dateutil==2.7.2
python-debian==0.1.27
python-systemd==231
python-telegram-bot==8.1.1
pytz==2018.3
PyWavelets==0.5.2
pyxdg==0.25
PyYAML==3.12
reportlab==3.3.0
requests==2.18.4
scikit-learn==0.17.1
scipy==1.0.1
service-identity==17.0.0
sessioninstaller==0.0.0
six==1.11.0
spicy==0.16.0
system-service==0.3
tabulate==0.8.2
tensorflow==1.3.0
tensorflow-tensorboard==0.1.7
Theano==1.0.1
Twisted==17.9.0
txaio==2.9.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.22
usb-creator==0.3.0
Werkzeug==0.14.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
zope.interface==4.4.3
